Force cricket back onto the screen
The previous attempt of ++/-- inc/dec wasn't strong enough to keep the
cricket from falling off because early in the jump the X velocity is
more than one, so it escapes the screen faster than it gets pushed back.
Instead of inc/dec I now just reset it to where the edge is supposed to
be, i.e. 0 on the left and screen width - cricket width on the right.