Prime cricket by holding down space
The time space was held down for "prime duration" is saved to the
Cricket object so that in the next tick, if the spacebar has been let go
it can use that value.  There are many ticks per second, which makes it
very sensitive, so I'm dividing it by the global "wait time".