Flip cricket sprite's direction to face the way it's jumping
You do this by scaling it -1 on the X axis and keeping Y the same.
Unfortunately after this it has been flipped about the origin, so while
0, 0 was previously top-left it's now top-right, so to compensate for
this I translate the sprite back to the right by the width of the
cricket image so that it stays in the right place visually.