Switch map levels by pressing the N key
Added to more example maps in the LDtk project, pressing N on the last 1
will wrap around to the first one.  After changing the level it's
automatically re-rendered because that's always based on the current
level.  The cricket jumping state and position is also reset to the
start position on the new level at the time of changing.