Change collision to use hitbox against each tile
Loops through all the tiles and autotiles for the level, considers them
as rectangles and checks whether they overlap with the rectangle hitbox
for the cricket.  Luckily the map is quite small so this goes fast.
Used a helper function for the looping because it's exactly the same for
both kinds of tiles.