Differentiate between passible and impassible tiles
The current collision response stops you from moving through any tiles.
Now there's a list of tiles you can't pass, for which this logic
applies, and for the rest (plants, mushrooms, etc.) it allows you to
somewhat move through them.  It still needs work but it's an improvement
on the previous strict collision logic.

Resolves #10