Make Android build work!!!
This took a surprisingly large amount of work but it works!  I ran into
several issue the process that needed to be worked around, and those
needed cleaning up, which is why I'm only committing this now despite it
working for more than a month already:

The whole project needed to be restructured because so far I was
building main.go in a binary just by `go build .` , but gomobile doesn't
want a main file, it wants a library file that it'll load itself.  So, I
had to split the project, renaming what was the main file before into
game.go and adding a new main package at bin/main.go which only does the
window setup, initiate the main Game object and start the game.  I then
copied that and made a similar version in mobile/mobile.go that fulfils
gomobile's requirements.

This meant all the files' packages had to be renamed and it was at this
point the compile complained about underscores in the package name, so I
renamed the whole project, on GitHub, locally, on itch.io etc. from
cr1ck_t to cr1ckt with no underscore.

I also ran into an issue with the ebitenrender I'd imported from the
ldtkgo repo because it assumes there is filesystem access and although
for Web it was enough to write two functions to tell it to use instead,
which used the embedded FS, for Android this was not the case and even
the very mention of ebitenutil.ImageFromFile caused the whole thing to
fail because instead of a runtime error, it errors at build time, as in
ebitenmobile that function doesn't even exist at all.  So I copied over
the whole example renderer, deleted those calls and replaced the 2
needed functions, similarly to before but from my own code.  This has
got it building and working now.

One last unresolved issue is that including the MP3 inside the game
binary makes the AAR file gigantic and also gives you just a black
screen when you start the game.  Removing the music solves the problem
and the game is playable on Android, not sure why that works.