Refactor collision detection logic into its own place
This collision detection came out of a pile of iterative improvements
that all worked one-by-one but mixed up collision detection and
collision response to the point where it was hard to improve either of
them at all.  Now that the collision response logic has its own place
I'm moving out collision detection into its own file.  So far it has a
Collides(*Game) function that returns which tile it was colliding with.
Just the act of separating this has reduced the number of nested levels
significantly and is hopefully a step towards something testable.