Move Cricket initialisation code to a common New function
The code for making a Cricket object at the start of the game resetting
it on level switch was almost exactly the same, the biggest difference
was the location for where the Cricket should be.  I've moved that code
to a NewCricket function that can be called from both places.