Add fruit icon and target level for Exit entities in maps
The fruit icon in the v2 tileset might not be intended for this purpose
but until there's another use for it I've set it as the icon for the
Exit entities.  I've also added an Integer value to the Exit entities,
called Target, that lets you specify which level the game should take
you to if you reach it.