Implement moving camera over map using ebitencamera
Instead of drawing the map and cricket directly to the screen, draw them
to a new, off-screen background image called bg, and let ebitencamera do
the work of moving about it and following the player, and blit the
result on the on screen image.

It has some things that can be optimised, like not recreating the
background image and redrawing all tiles each time Draw is called but at
least it's finally working.

Resolves #9