Make empty blackness map instead of nil map
If you landed on water or skipped to the next level the Reset function
would set the blackness map to nil and the next jump would try to set an
element in it to true.  This caused a crash because you can't add
elements to an uninitialised map, you need to make it first.  Fixed by
copying the map initialising "make" from the NewGame function.

Resolves #19