Concat cricket camera position to cricket direction
The position of the cricket on the screen is now handled by the camera's
GetTranslation function, so we shouldn't translate the Cricket's
internal GeoM any more.  It does have an up-to-date GeoM on the
direction the Cricket is facing though, so concat these two to get a
cricket facing the right direction and correctly positioned on the
screen.

I'm not sure why the order matters but concating the position to the
direction works but the other way round had bad results.

Resolves #21