Return early when hitting water or exit
If the Cricket hit one of these important tiles the Reset function would
run, setting a new cricket at the starting point, but when it returned,
the Update function continued and would lead to strange behaviour like
setting the cricket jump value back to false in a later logic branch.
Because when we hit one of these we need game logic for that level to
stop so that the Reset one an start, we return early from the Update
function in both of these cases.

Resolves #16