Copy ebiten camera and make it work on web
The ebiten-camera library uses ebiten.WindowSize() which only works on
Desktop but I want this game to work on web and mobile too.  I've copied
down the library (luckily it's a single file), included the author's
license at the top of the source code and made it work without
ebiten.WindowSize().  To do this, I pass it the set Width and Height of
the game as constructor arguments, save them in the camera object and
refer to those in all places where ebiten.WindowSize() was previously
being used.

Resolves #22