Use the map size instead of screen size to keep the Cricket on the map
The logic that makes you stay in the screen (before there was a camera)
stopped you if you reached the edge of the screen, not the edge of the
map, the difference didn't matter at the time because the screen size
was the same as the map size.  Now that there's a moving camera the
screen size is much smaller than the map size and this manifests in the
game as an invisible wall at the point where the screen size ends in the
map.  This is fixed by using the map size from the maps data instead.

Resolves #23