Use Cricket hitbox for detected edge of map collision
There was another invisible wall along the right side of the map if you
jumped into it because the width of the Cricket's animation frame was
used to calculate collision and that usually has some empty pixels.
Using the width of the Cricket's hitbox reduces that to a much nicer
size that doesn't look weird.