Clamp the Camera to the Map dimensions
Surely there is an easier way to do this with maths... ಠ_ಠ

Checking min and max for X and Y one by check whether the Cricket is
closer than half the screen to any edge of the map and in that case
position the screen halfway from that map edge.  Otherwise the default
camera logic of centring on the Cricket applies.