Add a minimum jump strength and set it to 2
Based on user feedback it's hard to tell if the game is doing anything
because when you tap the screen nothing happens, that's because you need
to hold it down for a short while to jump any real distance.

This makes so that even when you just quickly tap the screen it still
jumps a visible distance.  The minimum jump strength is configurable via
the config file as MinPrime.