Start the map with the Cricket falling out of the air
This is to avoid an interesting issue where starting on the ground has
the Cricket stuck in it and you need to jump with some force to get out,
which is confusing for new users.  It could probably be fixed but for
now it's easier just to start in the air.