Add test cases for OverlapsTiles
These test variations combinations of overlapping tiles of earth, water
and non-tile spaces to verify the current system works as expected.  It
also includes a case for matching a top-left tile first before other
tiles as both this solution and the next one are expected to start
checking the top-left corner first.