Explicitly name output and source file for EXE build
The project restructuring required to get the ebitenmobile build to work
for Android moved away the main.go file under bin/ which broke the
GitHub Actions build step which expects it to be in the top-dir when you
run `go build ./...` and also breaks the upload step which expects the
generated file to be called cr1ckt.exe.  I've ensured both these work
now by specifying them explicitly.  In future probably it would be more
convenient to add a Makefile for these build flags.