Replace camera implementation with unmodified ebiten-camera
The author of ebiten-camera has changed it to take the game's logical
width and height as parameters instead of calling ebiten.WindowSize()
directly.  This is great because it makes the lib portable to web where
ebiten.WindowSize() returns nil (0, 0).  So, the solution I did in #22
is not needed any more, hooray!