Add boilerplate file structure for MacOS app
To make a desktop-runnable Mac app with a name and icon you need to
generate icon.icns file (I used the PNG icon from the artwork directory)
and put and the built executable (with executable bit set) in the right
directory.  You also need to add an Info.plist file that mentions the
icon and executable by name.

I got the contents of the plist file from another app, deleted the parts
that looked irrelevant or I didn't know what they were for and rewrote
the rest to match my game and it works!  This is, as far as I can tell,
the most minimal set-up required to make it work, there's surely many
more options you can add but I haven't looked at them.