Extract test running step to before the builds run
Instead of repeating the tests on each platform I trust Ebiten to work
similarly on each platform.  Run the tests only once at the start, and
proceed with building on different platforms only if the tests pass.  I
picked Linux as the platform for testing because despite needing to
install additional dependencies it's still the fastest of all the
build OSes.