Draw map tiles once at game start instead of for each frame
There were two parts here:
- for every Update() regenerate the tiles (useful if you have animated
  tiles... I don't)
- for every Draw() draw those tiles onto the background, most of them
  aren't even shown on the screen

This was happening for every frame even though the map was huge,
additionally it'd draw to a new *ebiten.Image from scratch each frame
but all the tiles in the game are static and are known at the start of
the game, at most they would change when you go to the next level.  But
since there's only one level it's enough to draw this all onto a big
image once and then just blit the shown part onto the camera, it's much,
much faster!