Replace the mp3 audio with a shorter clipped Lo-Fi OGG version
The music for the game got a lot of positive comments but it also makes
the binary twice the size, so I cut off the part after the intro to make
it shorter and reencoded it as lower-quality (compressed) OGG.  The file
is now 924K instead of 9.9M, that's 10 times smaller!  Hooray for
smaller game sizes!