Play music as an infinite loop with an intro
You can add an infinite loop that loops back to a part in the song after
a specified amount of bytes used as an intro.  I used the cricket
chirping part as the intro and then after a few bars loop back to where
the drum beat starts, this sounds quite similar to the original song
(which was calm and repetitive) although slightly less varying but this
trick lets us use a much smaller file size.

The part in the song where there's a short silence before the first
kickdrum starts is at 53 seconds and the song is 1m25s long, so using
this ratio, I calculated the position in bytes of where the loop should
start, based on what Ebiten gives me as the number of bytes it's
counting as the song length.  That's where the magic number 10113930
comes from.