Add a window icon for the game
Using Ebiten's SetWindowIcon function I can set an icon that appears
when you alt-tab or look at the taskbar icon.  I used the same artwork
that's used as the Mac app icon.  This needs to be loaded from embedFS
so I used the code in files.go which is previously an unexported
function, so I might have to redesign that package a bit.