Rearrange the project to have Go files in their own place
They were starting to look a bit disorganised mixed up at the top level
with all the other general files, but mostly it got really annoying
always typing go run ./bin/main.go instead of just go run.

The game code needs to be in its own package for gomobile to work.  The
main.go file can't be in the same folder as the game library so instead
of putting main.go in a sub-folder I've moved all the other Go files to
their own folder.  Because these Go files aren't intended to be imported
from outside the game project, even if some of them look a bit like a
lib, they're not stable or production-ready so they've gone into the
internal folder.

The build seems to still work fine this way too, hooray!