Move the creation of the music stream to the file-handling code
I noticed the last line of the Android logs complained about an unclosed
file:

> W/System: A resource failed to call close.

Based on this I realised that one difference from my quickly written
code for loading music compared to other filetypes was that they close
the file after reading it and the music one doesn't.

I've reorganised the music loading code to delegate Vorbis decoding to
it and now it also closes the file after opening, hopefully this'll
resolve that issue.