Remove Linux testing dependencies from Web build
This "Install dependencies" step got copied from the Linux build
instructions, because both the Linux binary and the Web bundle are built
on Linux.  However, these dependencies aren't necessary for building the
app, only for testing it, so they can be safely removed from the build.
It should be slightly faster without them.