Initialise game in a goroutine in Android too
The code for the desktop version shoots off a goroutine for the game
initialisation once the required stuff (like setting Window title etc.)
is done because loading the media files, setting up the map and initial
game state can take a few seconds.  This lets the games first update
loop start sooner, and display the "Loading..." text until everything is
set up and the game can start.

I wasn't sure if goroutines would work in the Android version too but
I've tried it out and not only does it work but the problem of the game
being stuck at a blank screen at the start has also disappeared.  Great!