Include Android built artefacts in developer bundle
Included the APK for playtesters who want to try out the Android app.
I've also included the AAR file and JAR sources for anyone who's an
Android developer and wants to tweak the Android part of the code
without installing Go to build the lib, they can just pop this one in.