Update the download text for the dev bundle
So far it said you could get the latest Windows EXE by the download
link, and for other platforms a link to the section on how to build.
Since the bundle now also includes the other 3 platforms and additional
development files I've updated the text to say "latest development
bundle" and listed the contents of the bundle.