Connect Android lifecycle events to EbitenView handlers
I noticed that the game music kept playing when you put the app to the
background in Android.  EbitenView provides two functions suspendGame()
and resumeGame() that need to be hooked up to the app's main activity's
onPause() and onResume() event handlers respectively.  After hooking
them up the app responds to these states and everything behaves as you
would expect.