Make theme colours black so that there's no ugly flash during loading
While the app loads there is some default combination of ugly colours
from the base theme that only show if the device is slow enough.  Adding
a proper splash screen seems like too much effort considering the game
actually loads pretty fast, I just want to avoid this millisecond
flicker of ugly theme colours.  Setting those colours to black seems to
work well as a "poor man's splash screen", so it'll show a black screen
while loading the game.  In future putting the game icon or cricket in
the middle and removing it once the Ebiten Activity is ready would be a
cute touch.