Set the main activity as hardware accelerated
This is the one the game runs in and the default is false, so one would
expect that setting it to true would result in some cool performance
enhancement for the game.  In reality I'm not sure I can feel any
improvement, the animations still flicker a bit but I can't tell if it's
less now, at least it felt pretty responsive, so I'll leave this in
because it seems like an appropriate setting for a game.