Explicitly specify Android API 19 for ebitenmobile
If you leave out this parameter to ebitenmobile it'll use the default
API version (16?) which is unsupported (by gomobile?) and so the build
fails with this error:

    unsupported API version 16 (not in 19..33)

Setting it explicitly to 19 solves the issue. Incidentally 19 matches
the project's minSdkVersion which is nice, otherwise I'd have to bump
that too, because minSdkVersion can't be lower than this number.