Update go install syntax for using rsrc with go modules
Previously go get did the trick but now with go modules that command is
used for adding dependencies and you need to use go install with a
specified version (in this case @latest) in order to actually get the
tool installed to be able to run it.  So, updated the github workflow.